relocate metrics
